+++ src/main/java/com/donorapi/jpa/SlotsRepository.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SlotsRepository extends JpaRepository<Slot, Integer> {
+public interface SlotsRepository extends JpaRepository<Slot, Long> {
 
     @Query("SELECT s FROM Slot s WHERE s.hospital.hospitalId = :hospitalId")
-    List<Slot> findSlotsByHospitalId(@Param("hospitalId") Integer hospitalId);
+    List<Slot> findSlotsByHospitalId(@Param("hospitalId") Long hospitalId);
 
 
 }
